Finish bookingFeedback flow

+++ src/Business/Services/BookingFeedbackService.cs
@@ -45,36 +45,32 @@ namespace Business.Services
             await _bookingFeedbackRepository.CreateAsync(entity);
         }
 
-        public async Task UpdateByAdminAsync(Guid id, BookingFeedbackModel updateModel)
-        {
-            var entityToUpdate = await GetBookingFeedbackEntityAsync(id, updateModel);
-            await UpdateEntityAsync(entityToUpdate, updateModel);
-        }
-
-        public async Task UpdateByUserAsync(Guid feedbackId, BookingFeedbackModel updateModel, Guid userId)
+        public async Task UpdateAsync(Guid feedbackId, BookingFeedbackModel updateModel, Guid? userId)
         {
             if (feedbackId != updateModel.Id)
             {
-                throw new BadRequestException("Check data!");
+                throw new BadRequestException("Check identifiers!");
             }
 
             var entityToUpdate = await _bookingFeedbackRepository.GetAsync(feedbackId);
+
             if (entityToUpdate == null)
             {
-                throw new NotFoundException($"{nameof(updateModel)} with id = {feedbackId} not found.");
+                throw new NotFoundException("Booking feedback not found!");
             }
 
-            if (entityToUpdate.UserId != userId)
+            if (userId != null && userId != entityToUpdate.UserId)
             {
-                throw new NotAuthorizedException("You do not have permissions to do this action!");
+                throw new NotAuthorizedException("No permissions for this action!");
             }
 
             entityToUpdate.Comment = updateModel.Comment;
             entityToUpdate.Rating = updateModel.Rating;
+
             await _bookingFeedbackRepository.UpdateAsync(entityToUpdate);
         }
 
-        public async Task DeleteAsync(Guid bookingFeedbackId)
+        public async Task DeleteAsync(Guid bookingFeedbackId, Guid? userId)
         {
             var entityToDelete = await _bookingFeedbackRepository.GetAsync(bookingFeedbackId);
             if (entityToDelete == null)
@@ -82,29 +78,12 @@ namespace Business.Services
                 throw new NotFoundException($"{nameof(entityToDelete)} with id = {bookingFeedbackId} not found.");
             }
 
+            if (userId != null && userId != entityToDelete.UserId)
+            {
+                throw new NotAuthorizedException("No permissions for this action!");
+            }
+
             await _bookingFeedbackRepository.DeleteAsync(entityToDelete);
-        }
-        private async Task<BookingFeedbackEntity> GetBookingFeedbackEntityAsync(Guid feedbackId, BookingFeedbackModel updateModel)
-        {
-            if (feedbackId != updateModel.Id)
-            {
-                throw new BadRequestException("Check data!");
-            }
-
-            var entityToUpdate = await _bookingFeedbackRepository.GetAsync(feedbackId);
-            if (entityToUpdate == null)
-            {
-                throw new NotFoundException($"{nameof(updateModel)} with id = {feedbackId} not found.");
-            }
-
-            return entityToUpdate;
-        }
-
-        private async Task UpdateEntityAsync(BookingFeedbackEntity entityToUpdate, BookingFeedbackModel updateModel)
-        {
-            entityToUpdate.Comment = updateModel.Comment;
-            entityToUpdate.Rating = updateModel.Rating;
-            await _bookingFeedbackRepository.UpdateAsync(entityToUpdate);
         }
     }
 }
